create sections on save
Try to create sections on save. Sectionized HTML is currently not
saved because a HTML to XHTML JS converter is still needed. For getting
this working in all situations we need a HTML to XHTML JS converter
(similar to HTMLTidy5).
The sectionized HTML will be shown in the console for now.